agentic/adk/requirements.txt
google-adk==1.0.0
google-genai==1.14.0
google-cloud-aiplatform==1.93
python-dotenv>=1.0.0